pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "youtube-analytics-pipeline"
version = "0.1.0"
description = "YouTube Analytics Data Pipeline"
requires-python = ">=3.12"


# dependencies
dependencies = [
    "python-dotenv==1.0.0",
    "pydantic>=2.5.0,<3.0.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "google-cloud-bigquery>=3.23.1",
    "google-api-python-client>=2.100.0",
    "google-auth>=2.23.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
]


# optional dependencies
[project.optional-dependencies]

# pip install -e ".[extract]"
extract = [
    "psycopg2-binary==2.9.9",
    "isodate==0.6.1",
]

# pip install -e ".[orchestrate]"
orchestrate = [
    "prefect==3.0.0",
    "prefect-gcp==0.4.0",
    "dbt-core==1.7.0",
    "dbt-bigquery==1.7.0",
]

# pip install -e ".[serve]"
serve = [
    "streamlit>=1.32.0",
    "plotly>=5.20.0",
    "altair>=5.2.0",
]

# pip install -e ".[dev]"
dev = [
    "jupyter==1.0.0",
    "ipython==8.19.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

# pip install -e ".[all]"  →  cài tất cả
all = [
    "youtube-analytics-pipeline[extract]",
    "youtube-analytics-pipeline[orchestrate]",
    "youtube-analytics-pipeline[serve]",
    "youtube-analytics-pipeline[dev]",
]


# tool config
[tool.setuptools.packages.find]
where = ["."]
include = ["extract*", "orchestrate*", "serve*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "check_*.py"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]   # pycodestyle + pyflakes + isort
